requirements.txt
amqp==2.5.2
appdirs==1.4.4
asgiref==3.2.3
attrs==19.3.0
billiard==3.6.1.0
cached-property==1.5.1
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2
django-appconf==1.0.4
django-celery-beat==1.6.0
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-environ==0.4.5
django-mathfilters==1.0.0
django-qr-code==1.1.0
django-timezone-field==3.1
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
docopt==0.6.2
flower==0.9.5
html5lib==1.0.1
humanize==3.1.0
idna==2.8
importlib-metadata==1.5.0
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.7
lxml==4.5.1
MarkupSafe==1.1.1
num2words==0.5.10
numpy==1.18.0
pandas==0.25.3
Pillow==6.2.1
prometheus-client==0.8.0
psycopg2==2.8.4
PyPDF2==1.26.0
pyTelegramBotAPI==3.6.7
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
qrcode==6.1
redis==3.4.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
six==1.13.0
spyne==2.13.14b0
sqlparse==0.3.0
tornado==6.1
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zeep==3.4.0
zipp==2.1.0
